Ajouter le support de plusieurs langues

+++ aspi.sh
@@ -7,7 +7,7 @@
 #test si l'argument est vide
 if [ -z "$1" ]
   then
-    echo -e "\nSyntax: ./aspi.sh proxy port url user passwd \n\n"
+    echo -e "\nSyntax: ./aspi.sh proxy port url user passwd new_cookie\n\n"
     exit
 fi
 
@@ -19,22 +19,27 @@ log=$4
 pwd=$5
 cookies="cookies.txt"
 agent="Mozilla/5.0"
+new_cookie=$6
 
-rm -Rf $cookies
+if [ ! -z "$new_cookie" ]
+then
+    rm -Rf $cookies
+fi
 
 # authenticate and save cookies
-wget \
-    --user-agent="$agent" \
-    --save-cookies $cookies \
-    --keep-session-cookies \
-    --delete-after \
-    --post-data="log=$log&pwd=$pwd&testcookie=1" \
-    -q -e http_proxy="http://"$proxy":"$port"/" "$login_address"
+if [ ! -f "$cookies" ]
+then
+    wget \
+        --user-agent="$agent" \
+        --save-cookies $cookies \
+        --keep-session-cookies \
+        --delete-after \
+        --post-data="log=$log&pwd=$pwd&testcookie=1" \
+        -q -e http_proxy="http://"$proxy":"$port"/" "$login_address"
+fi
 
 # access home page with authenticated cookies
 wget \
     --user-agent="$agent" \
     --load-cookies $cookies \
     -qO- -e http_proxy="http://"$proxy":"$port"/" "$site"
-
-rm -Rf $cookies
